% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-helpers.R
\name{add_clock_time}
\alias{add_clock_time}
\title{Add clock times to a data frame of nodes}
\usage{
add_clock_time(df, df_rec)
}
\arguments{
\item{df}{A data frame created by one of the \code{gather} functions. Must
contain a \code{startTime} and \code{endTime} column that contain the seconds
since the start of the recording.}
}
\value{
A data frame with \code{startClockTime}, \code{endClockTime},
\code{startClockTimeLocal}, and \code{endClockTimeLocal} columns for each
recording.
}
\description{
The segment and block nodes only have startTime and endTime attributes that
indicate the time since the beginning of the corresponding recording (in
seconds). This function takes a data frame with recording information and
adds the corresponding clock time (UTC) and local time (adjusted for
timezone). Note that the local time is correctly adjusted but still stored
with timezone "UTC". See \code{\link{as_local_time}} for more information.
}
\keyword{internal}
