% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-helpers.R
\name{add_col_if_missing}
\alias{add_col_if_missing}
\title{Add default-value columns to a dataframe}
\usage{
add_col_if_missing(df, ...)
}
\arguments{
\item{df}{a dataframe}

\item{...}{key-value pairs. The lefthand side should be the name of the
desired column and the righthand should be the value to use as a default.}
}
\value{
the dataframe is returned. If any of columns named in \code{...} are
  missing, they are added to the dataframe.
}
\description{
Add default-value columns to a dataframe
}
\examples{
\dontrun{
add_col_if_missing(
  iris,
  Species = "col exists. this shouldn't do anything",
  Species2 = "col doesn't exist. this should be added")
}
}
\keyword{internal}
