% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{add_id_attrs}
\alias{add_id_attrs}
\title{Add recording id, block id, and segment id to all annotation nodes
These ids can later serve as foreign keys for join operations etc.}
\usage{
add_id_attrs(its_xml)
}
\description{
Add recording id, block id, and segment id to all annotation nodes
These ids can later serve as foreign keys for join operations etc.
}
\keyword{internal}
