% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-helpers.R
\name{add_its_id}
\alias{add_its_id}
\title{Add \code{.its} filename to a data frame produced by any of the
\code{gather_} functions}
\usage{
add_its_id(df, its_xml)
}
\arguments{
\item{df}{A data frame produced by one of the \code{gather_} functions}

\item{its_xml}{The \code{.its} xml tree corresponding to the data frame}
}
\value{
A data frame eith an additional column \code{its_file}, containing
the name of the corresponding \code{.its} file
}
\description{
Add \code{.its} filename to a data frame produced by any of the
\code{gather_} functions
}
\keyword{internal}
