% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-helpers.R
\name{as_local_time}
\alias{as_local_time}
\title{Use UTC time and timezone from \code{.its} file to calculate local time.}
\usage{
as_local_time(utc_time, local_tzs)
}
\arguments{
\item{utc_time}{A vector of POSIXct object to convert to local time}

\item{local_tzs}{A single local timezone to convert to or a vector of local
timezones with the same length as \code{utc_time}}
}
\description{
In .its files all timestamps are stores in UTC time the actual timezone is
stored separately. This function converts the timestamps back to local time.
Local times are stored with timezone UTC, regardless of the actual timezone.
This is done for two resons: (1) So that "time of the day" can be compared
across LENA recordings made in different timezones, without converting to a
common time zone first. (2) Because time objects with
different timezones cannot be stored in the same column of a data frame.
}
\keyword{internal}
