% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-helpers.R
\name{clean_its_time}
\alias{clean_its_time}
\title{Convert time format strings used in .its to numeric (seconds)}
\usage{
clean_its_time(x)
}
\description{
Convert time format strings used in .its to numeric (seconds)
}
