% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-helpers.R
\name{clean_time_cols}
\alias{clean_time_cols}
\title{Coerce all columns with LENA time strings to numeric (seconds)}
\usage{
clean_time_cols(df)
}
\arguments{
\item{df}{A data frame produced by any of the \code{gather_} functions.}
}
\value{
A data frame with all time columns converted to numeric
}
\description{
Coerce all columns with LENA time strings to numeric (seconds)
}
\keyword{internal}
