% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-helpers.R
\name{clean_tz_str}
\alias{clean_tz_str}
\title{Convert timezone string used in \code{.its} to \code{lubridate} compatible
timezone name.}
\usage{
clean_tz_str(tz_str)
}
\arguments{
\item{tz_str}{An \code{.its} timezone string}
}
\description{
The timezone format in the its files is not compatible with lubridate.
This function fixes this. See \code{OlsonNames()} for valid timezone names.
}
\keyword{internal}
