% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{gather_recordings}
\alias{gather_blocks}
\alias{gather_conversations}
\alias{gather_pauses}
\alias{gather_segments}
\alias{gather_ava_info}
\alias{gather_child_info}
\title{Extract nodes of a certain type from an '.its' file as data frame.}
\usage{
gather_recordings(its_xml)

gather_blocks(its_xml)

gather_conversations(its_xml)

gather_pauses(its_xml)

gather_segments(its_xml)

gather_ava_info(its_xml)

gather_child_info(its_xml)
}
\arguments{
\item{its_xml}{the xml tree of an .its file}
}
\description{
gather_recordings - Extract all recording nodes,  as data frame.
gather_blocks - Extract all block nodes (pauses and conversations) as data
frame.
gather_conversations - Extract all conversation nodes as data frame.
gather_pauses - Extract all pause nodes as data frame.
gather_segments - Extract all segment nodes as data frame.
#gather_ava_info - Extract Automatic Vocalization Assessment info as data
frame.
gather_child_info - Extract all child info as data frame.
}
