% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-helpers.R
\name{extract_its_filename}
\alias{extract_its_filename}
\title{Extract its filename from the \code{.its} xml tree}
\usage{
extract_its_filename(its_xml)
}
\arguments{
\item{its_xml}{An \code{.its} xml tree}
}
\value{
A character string, the name of the \code{.its} file as found in
the \code{.its} xml tree
}
\description{
Extract its filename from the \code{.its} xml tree
}
\keyword{internal}
