% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-helpers.R
\name{extract_tz}
\alias{extract_tz}
\title{Extract timezone of the recording from \code{.its}}
\usage{
extract_tz(its_xml)
}
\arguments{
\item{its_xml}{An \code{.its} xml tree}
}
\description{
Extract timezone of the recording from \code{.its}
}
\keyword{internal}
