% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{fetch_attr_path}
\alias{fetch_attr_path}
\title{Run xpath attribute query, storing results in a dataframe}
\usage{
fetch_attr_path(tree, path, label = NULL)
}
\arguments{
\item{tree}{an xml document loaded by xml2}

\item{path}{an xpath to an attribute value in a node}

\item{label}{an optional category label for the query}
}
\value{
a data_frame object with the columns XPath and Value. If a label was
  supplied, an additional Category column is included.
}
\description{
Run xpath attribute query, storing results in a dataframe
}
\keyword{internal}
