% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml-helpers.R
\name{fetch_attr_paths}
\alias{fetch_attr_paths}
\title{Run multiple xpath attribute queries, collecting results in a dataframe}
\usage{
fetch_attr_paths(tree, paths, labels = NULL)
}
\arguments{
\item{tree}{an xml document loaded by xml2}

\item{paths}{a vector of xpaths queries to run}

\item{labels}{an optional vector of category labels for each query}
}
\value{
a data_frame object with the columns XPath and Value. If labels were
  supplied, an additional Category column is included.
}
\description{
Run multiple xpath attribute queries, collecting results in a dataframe
}
\keyword{internal}
