% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{gather_ava_info}
\alias{gather_ava_info}
\title{Extract Automatic Vocalization Assessment info as data frame.}
\usage{
gather_ava_info(its_xml)
}
\arguments{
\item{its_xml}{the xml tree of an .its file}
}
\description{
Extract Automatic Vocalization Assessment info as data frame.
}
