% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{gather_child_info}
\alias{gather_child_info}
\title{Extract all child info as data frame.}
\usage{
gather_child_info(its_xml)
}
\arguments{
\item{its_xml}{the xml tree of an .its file}
}
\description{
Extract all child info as data frame.
}
