% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{gather_conversations}
\alias{gather_conversations}
\title{Extract all conversation nodes as data frame.}
\usage{
gather_conversations(its_xml)
}
\arguments{
\item{its_xml}{the xml tree of an .its file}
}
\description{
Extract all conversation nodes as data frame.
}
