% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{gather_path}
\alias{gather_path}
\title{Return nodes of a certain type from an '.its' file as data frame.}
\usage{
gather_path(its_xml, path, add_clock_time = TRUE)
}
\arguments{
\item{its_xml}{an .its file parsed with \code{\link{read_its_file}}}

\item{path}{A character string. The xpath query to run}

\item{add_clock_time}{Logical. If TRUE (default) colums \code{startClockTime},
\code{endClockTime}, \code{startClockTimeLocal}, and \code{endClockTimeLocal}
will be added to the data frame. They contain timestams corresponding to the
\code{startTime}, and \code{endTime} columns.}
}
\value{
A data frame containing the extracted nodes
}
\description{
Return nodes of a certain type from an '.its' file as data frame.
}
\keyword{internal}
