% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{gather_pauses}
\alias{gather_pauses}
\title{Extract all pause nodes as data frame.}
\usage{
gather_pauses(its_xml)
}
\arguments{
\item{its_xml}{the xml tree of an .its file}
}
\description{
Extract all pause nodes as data frame.
}
