% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{gather_speaker_transitions}
\alias{gather_speaker_transitions}
\alias{gather_speaker_transitions_from_segments}
\title{Gather speaker transitions}
\usage{
gather_speaker_transitions(its_xml, legal_transitions = NULL)

gather_speaker_transitions_from_segments(data_segments,
  legal_transitions = NULL)
}
\arguments{
\item{its_xml}{an .its file parsed with \code{\link{read_its_file}}}

\item{legal_transitions}{a character vector with transitions to keep. If
`NULL` (the default), only `c("MAN_CHN", "MAN_CHN", "FAN_CHN", "CHN_FAN")`
are used.}

\item{data_segments}{a dataframe produced by `gather_segments()`}
}
\value{
a dataframe with one row per segment. It contains the columns
  `transSpkr` (speaker transition as `previous_current`,) `transSegId`
  (segment IDs), and `transTime` (the time lag between the two segments).
}
\description{
Combs throughs speech segments and returns a dataframe of speaker
transition.
}
\keyword{internal}
