% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-helpers.R
\name{sort_gathered_columns}
\alias{sort_gathered_columns}
\title{Sort columns in data frames produced by any of the \code{gather_} functions}
\usage{
sort_gathered_columns(df)
}
\arguments{
\item{df}{A data frames produced by any of the \code{gather_} functions}
}
\value{
The data frame with columns in a different order.
}
\description{
Sort columns in data frames produced by any of the \code{gather_} functions
}
\keyword{internal}
