% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-helpers.R
\name{split_conversation_info}
\alias{split_conversation_info}
\title{Split conversationInfo column}
\usage{
split_conversation_info(segments_df)
}
\arguments{
\item{segment_df}{A data frame produced by \code{gather_segments()}}
}
\value{
A data frame with additional columns with conversation information
}
\description{
Some segment nodes contain a \code{conversationInfo} attribute with
additional information related to vocalization activity blocks and turns.
This function parses this information into separate columns.
}
\keyword{internal}
