% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml-helpers.R
\name{xml_path_to_df}
\alias{xml_path_to_df}
\title{Run xpath query and return attributes of the found nodes as data frame.}
\usage{
xml_path_to_df(its_xml, path, add_nodeType = FALSE)
}
\arguments{
\item{its_xml}{The xml tree returned by \code{read_its_file}}

\item{path}{A character string. The xpath query to run}

\item{add_node_type}{Logical. If TRUE a column with the node type
(name) of the extracted nodes is added to the data frame. If FALSE (default)
only the nodes attributes are returned as columns.}
}
\value{
A data frame. Rows correspond tho the nodes rturned from the query,
columns correspond to attributes of the returned nodes.
}
\description{
Run xpath query and return attributes of the found nodes as data frame.
}
\keyword{internal}
