% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-helpers.R
\name{clean_time_str}
\alias{clean_time_str}
\title{Convert \code{.its} time format strings to numeric (seconds)}
\usage{
clean_time_str(x)
}
\arguments{
\item{x}{An \code{.its} time string}
}
\value{
A numeric (time in seconds)
}
\description{
Convert \code{.its} time format strings to numeric (seconds)
}
\keyword{internal}
