% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-helpers.R
\name{clean_tz_str}
\alias{clean_tz_str}
\title{Make \code{.its} timezone string \code{lubridate} compatible}
\usage{
clean_tz_str(tz_str)
}
\arguments{
\item{tz_str}{An \code{.its} timezone string}
}
\description{
The GMT timezone string in the \code{.its} files is not compatible with
lubridate. This function fixes it. See \code{OlsonNames()} for valid
timezone names.
}
\keyword{internal}
