% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{extract}
\alias{gather_path}
\alias{gather_recordings}
\alias{gather_blocks}
\alias{gather_conversations}
\alias{gather_pauses}
\alias{gather_segments}
\alias{gather_ava_info}
\alias{gather_child_info}
\title{Extract nodes of a certain type from an '.its' file as data frame.}
\usage{
gather_path(its_xml, path, add_nodeType = FALSE)

gather_recordings(its_xml)

gather_blocks(its_xml)

gather_conversations(its_xml)

gather_pauses(its_xml)

gather_segments(its_xml)

gather_ava_info(its_xml)

gather_child_info(its_xml)
}
\arguments{
\item{its_xml}{an .its file parsed with \code{\link{read_its_file}}}
}
\value{
A data frame containing the extracted nodes.
}
\description{
Most \code{gather_} functions extract all nodes of a certain type from a
\code{.its} file.
\describe{
  \item{gather_recordings}{all recording nodes}
  \item{gather_blocks}{all block nodes (pauses and conversations)}
  \item{gather_conversations}{all conversation nodes}
  \item{gather_pauses}{all pause nodes}
  \item{gather_segments}{all segment nodes}
}
}
\details{
Other functions extract related nodes from the \code{.its}
\describe{
  \item{#gather_ava_info}{Automatic Vocalization Assessment info}
  \item{gather_child_info}{child info}
}
}
\keyword{internal}
